requirements.txt
numpy==1.24.3
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
plotly==5.15.0
streamlit==1.24.1
streamlit_lottie==0.0.5
